requirement.txt
fastapi
uvicorn
python-multipart
opencv-python-headless
ultralytics
tensorflow-cpu
numpy
Pillow